requirements.txt
requests
django
pytest
pytest-django
djangorestframework
djangorestframework-simplejwt
drf-nested-routers
asgiref==3.2.10
astroid==2.4.2
django-debug-toolbar==2.2
django-environ==0.4.5
flake8==3.8.3
gunicorn==20.0.4
lazy-object-proxy==1.4.3
Pillow==7.2.0
psycopg2-binary==2.8.5
pycodestyle==2.6.0
